requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.2
